[libc++] Extract destroy algorithms into separate headers
This patch separates the destroy algorithms from the primitive
construct_at and destroy_at operations, which are conceptually
not algorithms. This makes it easier to start using these destroy
algorithms from upcoming relocation facilities.
